gateexit-guardian-service/pyproject.toml
[tool.poetry]
name = "gateexit-guardian-service"
version = "0.1.0"
description = ""
authors = ["Patrick Kalkman <patrick@simpletechture.nl>"]
readme = "README.md"
packages = [{include = "gateexit_guardian_service"}]

[tool.poetry.dependencies]
python = "^3.11"
paho-mqtt = "^1.6.1"
loguru = "^0.7.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
